7891 Mi Alexander Felix README.md
This Is My Introduction and contact details. As I am a Cyber-JunkA.I. and So Wish To better the internet As a whole I do not Expect Compensation for any work on my GitHUB profile, However I am Homeless on the Streets and I wish to Create such AI marvels of work that you do not see an option but to forward my first email with a E-Transfer to be able to better feed my self and not look stupid with a kick ass hairstyle and clothes while typing on a computer at Mc Dicks for hours and not having any food to eat. I hope you will be Impressed and use so much everything that my Cyber-Maniac Mind provides for you that you will be willing to afford me a night at a motel or consult me to collaborate with you on your project. As a 2 TIME AVIATION VALEDICTORIAN I bring a complete different style to any projects I work on and a perspective that no computer engineer can provide for you anywhere in the world.